backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
openai>=1.40.0,<3.0.0
pydantic==2.9.2
httpx==0.27.2
psycopg2-binary==2.9.9
google-generativeai==0.7.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.0
